[refactor] createIndex 리팩토링
