+++ public/informationCrawlers/masterkey/crawler.js
@@ -2,8 +2,6 @@ const puppeteer = require("puppeteer"); // 크롤링을 위한 라이브러리
 const { MongoClient } = require("mongodb"); // MongoDB 사용을 위한 라이브러리
 const Redis = require("ioredis"); // Redis 사용을 위한 라이브러리
 const crypto = require("crypto"); // Hash데이터로 변경을 위한 라이브러리
-const { Client } = require("@elastic/elasticsearch"); // ElasticSearch사용을 위한
-const esClient = new Client({ node: "http://localhost:9200" });
 
 (async () => {
   console.time("Total Execution Time"); // 시간 측정 시작
@@ -50,139 +48,6 @@ const esClient = new Client({ node: "http://localhost:9200" });
     console.timeEnd("Total Execution Time");
   }
 })();
-
-async function createIndex() {
-  try {
-    // Check if the index already exists
-    const indexExists = await esClient.indices.exists({ index: "themes" });
-    if (indexExists) {
-      console.log("Index already exists. Deleting old index.");
-      await esClient.indices.delete({ index: "themes" });
-    }
-
-    // Create a new index with the specified settings and mappings
-    const response = await esClient.indices.create({
-      index: "themes",
-      body: {
-        settings: {
-          index: {
-            analysis: {
-              tokenizer: {
-                my_nori_tokenizer: {
-                  type: "nori_tokenizer",
-                  decompound_mode: "mixed",
-                  discard_punctuation: "false",
-                },
-                my_ngram_tokenizer: {
-                  type: "ngram",
-                  min_gram: 2,
-                  max_gram: 3,
-                },
-              },
-              // 이 필터가 없으면 문장에서 단어 사이에 공백까지도 토큰으로 만들기 때문에 검색 품질이 저하된다.
-              filter: {
-                stopwords: {
-                  type: "stop",
-                  stopwords: " ", // "" 토큰은 제거된다.
-                },
-              },
-              analyzer: {
-                my_nori_analyzer: {
-                  type: "custom",
-                  tokenizer: "my_nori_tokenizer",
-                  filter: ["lowercase", "stop", "trim", "stopwords", "nori_part_of_speech"],
-                  char_filter: ["html_strip"],
-                },
-                my_ngram_analyzer: {
-                  type: "custom",
-                  tokenizer: "my_ngram_tokenizer",
-                  filter: ["lowercase", "stop", "trim", "stopwords", "nori_part_of_speech"],
-                  char_filter: ["html_strip"],
-                },
-              },
-            },
-          },
-        },
-        mappings: {
-          properties: {
-            title: {
-              type: "text",
-              analyzer: "standard",
-              search_analyzer: "standard",
-              fields: {
-                nori: {
-                  type: "text",
-                  analyzer: "my_nori_analyzer",
-                  search_analyzer: "my_nori_analyzer",
-                },
-                ngram: {
-                  type: "text",
-                  analyzer: "my_ngram_analyzer",
-                  search_analyzer: "my_ngram_analyzer",
-                },
-              },
-            },
-            venue: {
-              type: "text",
-              analyzer: "standard",
-              search_analyzer: "standard",
-              fields: {
-                nori: {
-                  type: "text",
-                  analyzer: "my_nori_analyzer",
-                  search_analyzer: "my_nori_analyzer",
-                },
-                ngram: {
-                  type: "text",
-                  analyzer: "my_ngram_analyzer",
-                  search_analyzer: "my_ngram_analyzer",
-                },
-              },
-            },
-            explanation: {
-              type: "text",
-              analyzer: "standard",
-              search_analyzer: "standard",
-              fields: {
-                nori: {
-                  type: "text",
-                  analyzer: "my_nori_analyzer",
-                  search_analyzer: "my_nori_analyzer",
-                },
-                ngram: {
-                  type: "text",
-                  analyzer: "my_ngram_analyzer",
-                  search_analyzer: "my_ngram_analyzer",
-                },
-              },
-            },
-            genre: {
-              type: "text",
-              analyzer: "standard",
-              search_analyzer: "standard",
-              fields: {
-                nori: {
-                  type: "text",
-                  analyzer: "my_nori_analyzer",
-                  search_analyzer: "my_nori_analyzer",
-                },
-                ngram: {
-                  type: "text",
-                  analyzer: "my_ngram_analyzer",
-                  search_analyzer: "my_ngram_analyzer",
-                },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    console.log("Index created successfully:", response.body);
-  } catch (error) {
-    console.error("Error creating index:", error);
-  }
-}
 
 async function indexDataToElasticsearch(data) {
   const body = data.flatMap((doc) => [{ index: { _index: "themes" } }, doc]);
